Добавляет трип-контроллер
